redux not getting events but we be ok
