+++ react-app/src/components/HomePage/index.js
@@ -5,22 +5,45 @@ import { fetchAllGroups } from '../../store/groups';
 import { getCurrentUser } from '../../store/session';
 import EventGallery from '../EventGallery'
 import GroupGallery from '../GroupGallery'
+import SearchBar from '../SearchBar';
 import "./HomePage.css";
 
 const HomePage = () => {
-
-    const [ galleryToDisplay, setGalleryToDisplay ] = useState("event")
-    const [ eventListDisplayed, setEventListDisplayed ] = useState(true)
-    const dispatch = useDispatch();
+    let displayedGallery;
 
     const events = useSelector(reduxState => {
         return reduxState.events
     })
-
     const user = useSelector(reduxState => {
         return reduxState.session
     })
-    let displayedGallery;
+
+    const [query, setQuery] = useState('');
+    const [eventsShown, setEventsShown] = useState([...events]);
+    const [ galleryToDisplay, setGalleryToDisplay ] = useState("event")
+    const [ eventListDisplayed, setEventListDisplayed ] = useState(true)
+
+	let EventsRegex = new RegExp(query, 'i');
+	const filterEvents = () => {
+        if (query.length > 0) {
+            let eventName = [...events].filter((event) => EventsRegex.test(event.event_name));
+			let eventDescription = [...events].filter(event => EventsRegex.test(event.description))
+			let eventAddress = [...events].filter(event => EventsRegex.test(event.address))
+			let eventCity = [...events].filter(event => EventsRegex.test(event.city))
+			let eventState = [...events].filter(event => EventsRegex.test(event.state))
+			let eventZipcode = [...events].filter(event => EventsRegex.test(event.zip_code))
+            setEventsShown([...eventName, ...eventDescription, ...eventAddress, ...eventCity, ...eventState, ...eventZipcode,]);
+		} else if (query.length === 0) {
+            setEventsShown([...events]);
+		}
+	};
+
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        filterEvents();
+    }, [query, events]);
+
     useEffect(() => {
         dispatch(fetchAllEvents())
         dispatch(getCurrentUser())
@@ -31,16 +54,16 @@ const HomePage = () => {
             if (galleryToDisplay === "group"){
                return (
                 <>
-                    <GroupGallery user={user} type="all"/>            
+                    <GroupGallery user={user} type="all"/>
                     <br />
                     <GroupGallery user={user} type="user"/>
                 </>
-                ) 
-            }else {
+                )
+            } else {
                 return (
                 <>
                     <EventGallery />
-                </> 
+                </>
                 )
             }
         }
@@ -60,11 +83,17 @@ const HomePage = () => {
     return (
         <div className="home-page">
             <div className="home-page_header">
+                <div className="searchbar">
+                    <SearchBar
+                        query={query}
+                        setQuery={setQuery}
+                    />
+                </div>
                 <button className="toggle" id="group-toggle" hidden={!eventListDisplayed} onClick={()=>handleToggleClick()}>Groups</button>
                 <button className="toggle" id="event-toggle" hidden={eventListDisplayed} onClick={()=>handleToggleClick()}>Calendar</button>
             </div>
             <div className="home-page_body">
-                {displayedGallery}
+                <EventGallery />
             </div>
             <div className="home-page_sidebar">
             </div>
